requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
Flask-Cors
Flask-Migrate
Flask-JWT-Extended
Flask-BCrypt
SQLAlchemy
email-validator
python-slugify
python-dotenv
pytest